user-product/api/requirements.txt
fastapi==0.67.0
pandas==1.3.3
pydantic==1.8.2
pytest==6.2.5
python-dotenv==0.19.1
SQLAlchemy==1.4.25
starlette==0.14.2
mysql-connector-python==8.0.26
uvicorn
pymysql
